Fix availableFrom fading, today reactivity, and extract drag/drop logic
- availableFrom was stored and displayed but never actually faded future tasks;
  adds isNotYetAvailable() to useTasks and applies .not-yet-available opacity
  class in TaskCard so the feature claim matches the implementation
- today was a computed with no reactive dependencies so Vue cached it on first
  render and it never updated after midnight; replaces with a module-level ref
  polled every 60 s and shared via useTasks()
- extracts all HTML5 + touch drag logic from TaskCard into useTaskDrag() in
  useDragDrop.js, cutting TaskCard script from ~200 to ~90 lines
- adds a short AI involvement / human review section directly in README rather
  than relying solely on the wiki
