+++ tests/unit/task-editing/components.test.js
@@ -21,6 +21,9 @@ vi.mock('../../../src/composables/useTasks.js', () => ({
     updateTask: mockUpdateTask,
     moveTask: mockMoveTask,
     isOverCapacity: mockIsOverCapacity,
+    isNotYetAvailable: vi.fn(() => false),
+    today: { value: new Date().toISOString().slice(0, 10) },
+    moveTaskToColumn: vi.fn(),
     addSubtask: mockAddSubtask,
     deleteSubtask: mockDeleteSubtask,
     toggleSubtask: mockToggleSubtask,
